[MOV] calendar: move activity tests in their own file
Just to have them available in a separate file, easing future changes
and updates.

This notably highlights an issue with 'microsoft_calendar' (see
comments in tests). To be discussed and probably fixed accordingly.

Task-4813092: [calendar] Synchronize activity date with meeting
