+++ addons/calendar/tests/test_calendar_activity.py
@@ -1,4 +1,4 @@
-from datetime import datetime, date
+from datetime import datetime, date, timedelta
 from freezegun import freeze_time
 
 from odoo.addons.mail.tests.common_activity import ActivityScheduleCase
@@ -7,6 +7,154 @@ from odoo.tests import tagged, users
 
 @tagged('mail_activity_mixin', 'mail_activity', 'post_install', '-at_install')
 class TestCalendarActivity(ActivityScheduleCase):
+
+    def test_event_activity(self):
+        # ensure meeting activity type exists
+        meeting_act_type = self.env['mail.activity.type'].search([('category', '=', 'meeting')], limit=1)
+        if not meeting_act_type:
+            meeting_act_type = self.env['mail.activity.type'].create({
+                'name': 'Meeting Test',
+                'category': 'meeting',
+            })
+
+        # have a test model inheriting from activities
+        test_record = self.env['res.partner'].create({
+            'name': 'Test',
+        })
+        now = datetime.now()
+
+        # test with escaping
+        test_name, test_description, test_description2 = 'Test-Meeting', 'Test-Description', 'Test & <br> Description'
+        test_note, test_note2 = '<p>Test-Description</p>', '<p>Test &amp; <br> Description</p>'
+
+        # create using default_* keys
+        test_event = self.env['calendar.event'].with_user(self.user_employee).with_context(
+            default_res_model=test_record._name,
+            default_res_id=test_record.id,
+        ).create({
+            'name': test_name,
+            'description': test_description,
+            'start': now + timedelta(days=-1),
+            'stop': now + timedelta(hours=2),
+            'user_id': self.env.user.id,
+        })
+        self.assertEqual(test_event.res_model, test_record._name)
+        self.assertEqual(test_event.res_id, test_record.id)
+        self.assertEqual(len(test_record.activity_ids), 1)
+        self.assertEqual(test_record.activity_ids.summary, test_name)
+        self.assertEqual(test_record.activity_ids.note, test_note)
+        self.assertEqual(test_record.activity_ids.user_id, self.env.user)
+        self.assertEqual(test_record.activity_ids.date_deadline, (now + timedelta(days=-1)).date())
+
+        # updating event should update activity
+        test_event.write({
+            'name': '%s2' % test_name,
+            'description': test_description2,
+            'start': now + timedelta(days=-2),
+            'user_id': self.user_employee.id,
+        })
+        # TDE FIXME: currently failing if microsoft_calendar is installed as it
+        # creates a copy of the event (organizer change) and therefore copies
+        # activities also, see '_recreate_event_different_organizer'. Waiting
+        # proper investigation of behavior by calendar people.
+        if len(test_record.activity_ids) == 2:
+            activity = test_record.activity_ids[1]
+        else:
+            activity = test_record.activity_ids[0]
+        self.assertEqual(activity.summary, '%s2' % test_name)
+        self.assertEqual(activity.note, test_note2)
+        self.assertEqual(activity.user_id, self.user_employee)
+        self.assertEqual(activity.date_deadline, (now + timedelta(days=-2)).date())
+
+        # deleting meeting should delete its activity
+        test_record.activity_ids.unlink()
+        self.assertEqual(self.env['calendar.event'], self.env['calendar.event'].search([('name', '=', test_name)]))
+
+        # create using active_model keys
+        test_event = self.env['calendar.event'].with_user(self.user_employee).with_context(
+            active_model=test_record._name,
+            active_id=test_record.id,
+        ).create({
+            'name': test_name,
+            'description': test_description,
+            'start': now + timedelta(days=-1),
+            'stop': now + timedelta(hours=2),
+            'user_id': self.env.user.id,
+        })
+        self.assertEqual(test_event.res_model, test_record._name)
+        self.assertEqual(test_event.res_id, test_record.id)
+        self.assertEqual(len(test_record.activity_ids), 1)
+
+    def test_event_activity_multiple_meetings(self):
+        # Creating multiple meetings from an activity creates additional activities
+        # ensure meeting activity type exists
+        meeting_act_type = self.env['mail.activity.type'].search([('category', '=', 'meeting')], limit=1)
+        if not meeting_act_type:
+            meeting_act_type = self.env['mail.activity.type'].create({
+                'name': 'Meeting Test',
+                'category': 'meeting',
+            })
+
+        # have a test model inheriting from activities
+        test_record = self.env['res.partner'].create({
+            'name': 'Test',
+        })
+
+        activity_id = self.env['mail.activity'].create({
+            'summary': 'Meeting with partner',
+            'activity_type_id': meeting_act_type.id,
+            'res_model_id': self.env['ir.model']._get_id('res.partner'),
+            'res_id': test_record.id,
+        })
+
+        calendar_action = activity_id.with_context(default_res_model='res.partner', default_res_id=test_record.id).action_create_calendar_event()
+        event_1 = self.env['calendar.event'].with_context(calendar_action['context']).create({
+            'name': 'Meeting 1',
+            'start': datetime(2025, 3, 10, 17),
+            'stop': datetime(2025, 3, 10, 22),
+        })
+
+        self.assertEqual(event_1.activity_ids, activity_id)
+
+        total_activities = self.env['mail.activity'].search_count(domain=[])
+
+        event_2 = self.env['calendar.event'].with_context(calendar_action['context']).create({
+            'name': 'Meeting 2',
+            'start': datetime(2025, 3, 11, 17),
+            'stop': datetime(2025, 3, 11, 22),
+        })
+        self.assertEqual(event_1.activity_ids, activity_id, "Event 1's activity should still be the first activity")
+        self.assertEqual(activity_id.calendar_event_id, event_1, "The first activity's event should still be event 1")
+        self.assertEqual(total_activities + 1, self.env['mail.activity'].search_count(domain=[]), "1 more activity record should have been created (by event 2)")
+        self.assertNotEqual(event_2.activity_ids, activity_id, "Event 2's activity should not be the first activity")
+        self.assertEqual(event_2.activity_ids.activity_type_id, activity_id.activity_type_id, "Event 2's activity should be the same activity type as the first activity")
+        self.assertEqual(test_record.activity_ids, activity_id | event_2.activity_ids, "Resource record should now have both activities")
+
+    def test_event_activity_user_sync(self):
+        # ensure phonecall activity type exists
+        activty_type = self.env['mail.activity.type'].create({
+            'name': 'Call',
+            'category': 'phonecall'
+        })
+        activity = self.env['mail.activity'].create({
+            'summary': 'Call with Demo',
+            'activity_type_id': activty_type.id,
+            'note': 'Schedule call with Admin',
+            'res_model_id': self.env['ir.model']._get_id('res.partner'),
+            'res_id': self.env['res.partner'].create({'name': 'Test Partner'}).id,
+            'user_id': self.user_employee.id,
+        })
+        action_context = activity.action_create_calendar_event().get('context', {})
+        event_from_activity = self.env['calendar.event'].with_context(action_context).create({
+            'start': '2022-07-27 14:30:00',
+            'stop': '2022-07-27 16:30:00',
+        })
+        # Check that assignation of the activity hasn't changed, and event is having
+        # correct values set in attendee and organizer related fields
+        self.assertEqual(activity.user_id, self.user_employee)
+        self.assertEqual(event_from_activity.partner_ids, activity.user_id.partner_id)
+        self.assertEqual(event_from_activity.attendee_ids.partner_id, activity.user_id.partner_id)
+        self.assertEqual(event_from_activity.user_id, activity.user_id)
 
     @users('employee')
     def test_synchronize_activity_timezone(self):
